+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All Notable changes to `Slytherin` will be documented in this file
 
+## [Unreleased](https://github.com/rougin/slytherin/compare/v0.6.0...HEAD)
+
+### Removed
+- `HttpKernelInterface`
+
 ## [0.6.0](https://github.com/rougin/slytherin/compare/v0.5.0...v0.6.0)
 
 ### Added