+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All Notable changes to `Slytherin` will be documented in this file
 
 ## [Unreleased](https://github.com/rougin/slytherin/compare/v0.6.0...HEAD)
 
+### Added
+- HTTP method spoofing
+
+### Changed
+- Version of `filp/whoops`
+
 ### Removed
 - `HttpKernelInterface`